code/requirements.txt
faust-streaming==0.11.3
kafka-python==2.3.0
influxdb-client
